+++ packages/web/app/components/PatientAllergiesWarning.jsx
